backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
firebase-admin>=6.4.0
google-generativeai>=0.5.0
Pillow>=10.0.0
imagehash>=4.3.1
requests>=2.31.0
reportlab>=4.1.0
python-dotenv>=1.0.0

streamlit>=1.32.0
